Moves reusable components to its own file
